refactor(app): Move frontend to pmweb module
- Split requirements.txt in two
- flake8: Allow longer lines
