+++ config.py
@@ -23,4 +23,3 @@ try:
     from local_config import *
 except ImportError:
     pass
-
