+++ docker-compose.yml
@@ -34,7 +34,7 @@ services:
             - MONGO_PORT=27017
         command: python3 run.py
 
-    test:
+    testcam:
         extends:
             file: docker-compose.yml
             service: printermood
@@ -43,7 +43,7 @@ services:
         environment:
             - MONGO_HOST=mongo
             - MONGO_PORT=27017
-        command: nosetests
+        command: nosetests printermood
 
     # External libs
     mongo:
