+++ pmweb/__init__.py
@@ -1,0 +1,8 @@
+from flask import Flask
+from flask_socketio import SocketIO
+
+app = Flask(__name__)
+app.config.from_object('config')
+socketio = SocketIO(app)
+
+import pmweb.views  # flake8: noqa
