+++ printermood/utils.py
@@ -6,13 +6,6 @@ class TimeoutException(Exception):
     pass
 
 
-def top_emotion(emotions):
-    if type(emotions) != dict or len(emotions) == 0:
-        return None
-
-    return sorted(emotions.items(), key=lambda x: x[1], reverse=True)[0]
-
-
 def are_rectangles_overlapping(rect1, rect2):
     """Returns `True` if two rectangles are overlapping.
