+++ requirements-web.txt
@@ -1,0 +1,9 @@
+IndicoIo==0.13.2
+nose==1.3.7
+Pillow==3.1.1
+requests==2.9.1
+Flask==0.10.1
+Flask-PyMongo==0.4.1
+Flask-SocketIO==2.2
+Flask-WTF==0.12
+python-magic==0.4.10
